Update key.dart
Add the ability to construct a Key from the base64 version. Most of the time, the Key is stored in the form of a string. It becomes necessary that a Key can be constructed from the said string, otherwise there's no use of storing the string in the first place.